fix issue #171 Binding transport to IPv6 address causes IllegalFormatConversionException
When starting Scalecube and one of the addresses only has an IPv6
address (for various reasons, in my case the DHCP didn't complete, but
we could really be using only an IPv6 network) then Addressing uses
NetworkInterface.getInetAddresses() to get an Inet6Address which looks
something like this: /fe80:0:0:0:8cf6:f5c8:c946:2c30%eno1.

Then when MembershipProtocol constructor tries to use the address to
create a "name format" for ThreadFactoryBuilder, the % in the address
causes the thread factory builder to crash with a
java.util.IllegalFormatConversionException when it tries to create
thread names using String.format().